+++ x/finality/keeper/power_dist_change_test.go
@@ -431,6 +431,8 @@ func FuzzBTCDelegationEvents_NoPreApproval(f *testing.F) {
 		stakingValue := int64(2 * 10e8)
 		delSK, _, err := datagen.GenRandomBTCKeyPair(r)
 		h.NoError(err)
+
+		stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
 		stakingTxHash, msgCreateBTCDel, actualDel, _, _, _, err := h.CreateDelegation(
 			r,
 			delSK,
@@ -454,13 +456,8 @@ func FuzzBTCDelegationEvents_NoPreApproval(f *testing.F) {
 		events := h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, btcTip.Height, btcTip.Height)
 		require.Len(t, events, 0)
 
-		btckptParams := btccKeeper.GetParams(h.Ctx)
-		stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
-
-		minUnbondingTime := types.MinimumUnbondingTime(&stakingParams, &btckptParams)
-
-		// the BTC delegation will be unbonded at end height - max(w, min_unbonding_time)
-		unbondedHeight := actualDel.EndHeight - minUnbondingTime
+		// the BTC delegation will be unbonded at end height - min_unbonding_time
+		unbondedHeight := actualDel.EndHeight - stakingParams.MinUnbondingTimeBlocks
 		events = h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, unbondedHeight, unbondedHeight)
 		require.Len(t, events, 1)
 		btcDelStateUpdate := events[0].GetBtcDelStateUpdate()
@@ -606,13 +603,8 @@ func FuzzBTCDelegationEvents_WithPreApproval(f *testing.F) {
 		require.Equal(t, stakingTxHash, btcDelStateUpdate.StakingTxHash)
 		require.Equal(t, types.BTCDelegationStatus_ACTIVE, btcDelStateUpdate.NewState)
 
-		btckptParams := btccKeeper.GetParams(h.Ctx)
-		stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
-
-		minUnbondingTime := types.MinimumUnbondingTime(&stakingParams, &btckptParams)
-
-		// the BTC delegation will be unbonded at end height - max(w, min_unbonding_time)
-		unbondedHeight := activatedDel.EndHeight - minUnbondingTime
+		// the BTC delegation will be unbonded at end height - minUnbondingTime
+		unbondedHeight := activatedDel.EndHeight - h.BTCStakingKeeper.GetParams(h.Ctx).MinUnbondingTimeBlocks
 		events = h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, unbondedHeight, unbondedHeight)
 		require.Len(t, events, 1)
 		btcDelStateUpdate = events[0].GetBtcDelStateUpdate()
@@ -678,6 +670,7 @@ func TestDoNotGenerateDuplicateEventsAfterHavingCovenantQuorum(t *testing.T) {
 	stakingValue := int64(2 * 10e8)
 	delSK, _, err := datagen.GenRandomBTCKeyPair(r)
 	h.NoError(err)
+	stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
 	expectedStakingTxHash, msgCreateBTCDel, actualDel, _, _, _, err := h.CreateDelegation(
 		r,
 		delSK,
@@ -693,20 +686,15 @@ func TestDoNotGenerateDuplicateEventsAfterHavingCovenantQuorum(t *testing.T) {
 	h.NoError(err)
 	/*
 		at this point, there should be 1 event that BTC delegation
-		will become expired at end height - w
+		will become expired at end height - min_unbonding_time
 	*/
 	// there exists no event at the current BTC tip
 	btcTip := btclcKeeper.GetTipInfo(h.Ctx)
 	events := h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, btcTip.Height, btcTip.Height)
 	require.Len(t, events, 0)
 
-	btckptParams := btccKeeper.GetParams(h.Ctx)
-	stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
-
-	minUnbondingTime := types.MinimumUnbondingTime(&stakingParams, &btckptParams)
-
-	// the BTC delegation will be unbonded at end height - max(w, min_unbonding_time)
-	unbondedHeight := actualDel.EndHeight - minUnbondingTime
+	// the BTC delegation will be unbonded at end height - min_unbonding_time
+	unbondedHeight := actualDel.EndHeight - stakingParams.MinUnbondingTimeBlocks
 	events = h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, unbondedHeight, unbondedHeight)
 	require.Len(t, events, 1)
 	btcDelStateUpdate := events[0].GetBtcDelStateUpdate()